+++ packages/beachball/src/publish/publishToRegistry.ts
@@ -1,75 +1,76 @@
+import _ from 'lodash';
+import path from 'path';
 import { performBump } from '../bump/performBump';
 import { BumpInfo } from '../types/BumpInfo';
 import { BeachballOptions } from '../types/BeachballOptions';
 import { packagePublish } from '../packageManager/packagePublish';
 import { validatePackageVersions } from './validatePackageVersions';
 import { displayManualRecovery } from './displayManualRecovery';
-import _ from 'lodash';
-import path from 'path';
+import { toposortPackages } from './toposortPackages';
+import { shouldPublishPackage } from './shouldPublishPackage';
+import { validatePackageDependencies } from './validatePackageDependencies';
 
 export async function publishToRegistry(originalBumpInfo: BumpInfo, options: BeachballOptions) {
   const { registry, tag, token, access, timeout } = options;
   const bumpInfo = _.cloneDeep(originalBumpInfo);
-  const { modifiedPackages, newPackages } = bumpInfo;
-  const prepublish = options.hooks?.prepublish;
+  const { modifiedPackages, newPackages, packageInfos } = bumpInfo;
+  const prepublishHook = options.hooks?.prepublish;
 
   await performBump(bumpInfo, options);
 
   const succeededPackages = new Set<string>();
 
+  let invalid = false;
   if (!validatePackageVersions(bumpInfo, registry)) {
     displayManualRecovery(bumpInfo, succeededPackages);
+    invalid = true;
+  } else if (!validatePackageDependencies(bumpInfo)) {
+    invalid = true;
+  }
+
+  if (invalid) {
     console.error('No packages have been published');
     process.exit(1);
   }
 
-  [...modifiedPackages, ...newPackages].forEach(pkg => {
-    if (!bumpInfo.scopedPackages.has(pkg)) {
-      console.log(`Skipping publish for out-of-scope package ${pkg}`);
+  const packagesToPublish = toposortPackages([...modifiedPackages, ...newPackages], packageInfos);
+
+  packagesToPublish.forEach(pkg => {
+    const { publish, reasonToSkip } = shouldPublishPackage(bumpInfo, pkg);
+    if (!publish) {
+      console.log(`Skipping publish - ${reasonToSkip}`);
       return;
     }
 
     const packageInfo = bumpInfo.packageInfos[pkg];
-    const changeType = bumpInfo.packageChangeTypes[pkg];
-    if (changeType === 'none') {
-      return;
+    console.log(`Publishing - ${packageInfo.name}@${packageInfo.version}`);
+
+    // run the prepublish hook once, after version bumping but before actually executing npm publish (with retries)
+    if (prepublishHook) {
+      prepublishHook(path.dirname(packageInfo.packageJsonPath), packageInfo.name, packageInfo.version);
     }
-    if (!packageInfo.private) {
-      console.log(`Publishing - ${packageInfo.name}@${packageInfo.version}`);
 
-      let result;
-      let retries = 0;
+    let result;
+    let retries = 0;
 
-      // run the prepublish hook once, after version bumping but before actually executing npm publish (with retries)
-      if (prepublish) {
-        prepublish(path.dirname(packageInfo.packageJsonPath), packageInfo.name, packageInfo.version);
-      }
+    do {
+      result = packagePublish(packageInfo, registry, token, tag, access, timeout);
 
-      do {
-        result = packagePublish(packageInfo, registry, token, tag, access, timeout);
+      if (result.success) {
+        console.log('Published!');
+        succeededPackages.add(pkg);
+        return;
+      } else {
+        retries++;
 
-        if (result.success) {
-          console.log('Published!');
-          succeededPackages.add(pkg);
-          return;
-        } else {
-          retries++;
-
-          if (retries <= options.retries) {
-            console.log(`Published failed, retrying... (${retries}/${options.retries})`);
-          }
+        if (retries <= options.retries) {
+          console.log(`Published failed, retrying... (${retries}/${options.retries})`);
         }
-      } while (retries <= options.retries);
+      }
+    } while (retries <= options.retries);
 
-      displayManualRecovery(bumpInfo, succeededPackages);
-      console.error(result.stderr);
-      throw new Error('Error publishing, refer to the previous error messages for recovery instructions');
-    } else {
-      console.warn(
-        `Skipping publish of ${packageInfo.name} since it is marked private. Version has been bumped to ${packageInfo.version}`
-      );
-    }
+    displayManualRecovery(bumpInfo, succeededPackages);
+    console.error(result.stderr);
+    throw new Error('Error publishing, refer to the previous error messages for recovery instructions');
   });
-
-  return;
 }